attackstyles: Rename Attack Indicators plugin to Attack Styles
